backend/requirements.txt
Flask
Flask-Cors
python-dotenv
PyJWT
bcrypt
db-sqlite3
stripe